+++ R/geneticAlgorithm.R
@@ -248,7 +248,7 @@ GAfit <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "
         break
       } else if(i1 == nattempts) {
         if(length(G) == 0) {
-          stop("Failed to create initial population with good enough individuals. Check if there is something weird in the data or in the provided constraints? You can also set up the initial population manually using the argument 'initpop' of the function 'GAfit'.")
+          stop("Failed to create initial population with good enough individuals. Scaling the individual series so that the AR coefficients (of a VAR model) will not be very large (preferably less than one) should solve the problem. If needed, another package may be used to fit linear VARs so see which scalings produce relatively small AR coefficient estimates.")
         } else {
           G <- G[,sample.int(ncol(G), size=popsize, replace=TRUE)]
         }
@@ -263,7 +263,11 @@ GAfit <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "
       if(parametrization == "intercept") {
         ind <- change_parametrization(p=p, M=M, d=d, params=ind, constraints=constraints, structural_pars=structural_pars, change_to="mean")
       }
-      initpop[[i1]] <- sort_components(p=p, M=M, d=d, params=ind, structural_pars=structural_pars)
+      if(is.null(constraints)) {
+        initpop[[i1]] <- sort_components(p=p, M=M, d=d, params=ind, structural_pars=structural_pars)
+      } else {
+        initpop[[i1]] <- ind
+      }
     }
     G <- replicate(popsize, initpop[[sample.int(length(initpop), size=1)]])
   }